refactor(functions): move twitter error file
