--- packages/functions/src/utils/error.ts
@@ -1,8 +1,0 @@
-export interface TwitterClientErrorData {
-  code: number;
-  message: string;
-}
-
-export const twitterClientErrorHandler = (errors: TwitterClientErrorData[]): { errors: TwitterClientErrorData[] } => {
-  return { errors };
-};
